backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
pandas>=2.0.0
google-generativeai>=0.5.2
